requirements.txt
streamlit
numpy
matplotlib
pillow
opencv-python
tensorflow
scikit-image
qiskit
qiskit-aer
pyngrok
Pylance